pyproject.toml
[project]
name = "test-jwt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django~=4.2",
    "djangorestframework~=3.15",
    "djangorestframework-simplejwt~=5.4.0",
    "drf-spectacular~=0.28.0",
    "pytest~=8.3",
    "pytest-django~=4.9.0",
]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"
